+++ MinecraftWriter/net/morbz/minecraft/blocks/DoorBlock.java
@@ -76,8 +76,8 @@ public class DoorBlock implements IBlock {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public byte getBlockId() {
-		return (byte)material.getValue();
+	public int getBlockId() {
+		return material.getValue();
 	}
 	
 	/**
