+++ MinecraftWriter/net/therealvira/minecraft/blocks/SimpleBlockAdvanced.java
@@ -1,8 +1,14 @@
 package net.therealvira.minecraft.blocks;
 
 import net.morbz.minecraft.blocks.IBlock;
+import net.morbz.minecraft.blocks.Material;
 import net.morbz.minecraft.blocks.RedstoneTorchBlock;
+import net.morbz.minecraft.blocks.SimpleBlock;
 import net.morbz.minecraft.blocks.states.Facing5State;
+
+import java.lang.reflect.Field;
+import java.util.Arrays;
+import java.util.HashMap;
 
 public enum SimpleBlockAdvanced implements IBlock {
     STANDING_REDSTONE_TORCH(new RedstoneTorchBlock(true, Facing5State.UP)),
@@ -15,7 +21,7 @@ public enum SimpleBlockAdvanced implements IBlock {
     }
 
     @Override
-    public byte getBlockId() {
+    public int getBlockId() {
         return block.getBlockId();
     }
